+++ commands/mHelper.js
@@ -1,7 +1,7 @@
 var capitalize = require('./functions');
 module.exports = function (vscode, fs, path, pathdir) {
     vscode.window.showInputBox({
-        prompt: "name of helper",
+        prompt: "Enter name of helper",
         placeHolder: "Enter helper file"
     }).then(function (val) {
         if (val.length == 0) {
@@ -12,10 +12,7 @@ module.exports = function (vscode, fs, path, pathdir) {
             fs.access(pathfile, function (err) {
                 if (!err) {
                     vscode.window.showWarningMessage("Helper file name already exists!");
-
                 } else {
-
-
                     fs.open(pathfile, "w+", function (err, fd) {
                         if (err) throw err;
                         fs.writeFileSync(fd, `<?php 
@@ -25,29 +22,17 @@ function secureFunction(){
 }
 
 
-/* End of file ` + capitalize.lowercase(val) + `.php and path  /application/helpers/` + val + `_helper.php */
-
+/* End of file ${capitalize.lowercase(val)}.php and path ${pathfile.replace(pathdir,'')} */
 `);
                         fs.close(fd);
                         var openPath = vscode.Uri.file(pathfile); //A request file path
-
                         vscode.workspace.openTextDocument(openPath).then(function (val) {
                             vscode.window.showTextDocument(val);
-
                         });
-
                     });
                     vscode.window.showInformationMessage('Helper created successfully! ');
-
                 }
-
-
             });
-
-
         }
-
     });
-
-
 }