+++ commands/mSimpleAuth.js
@@ -2,20 +2,15 @@ var capitalize = require('./functions');
 module.exports = function (vscode, fs, path, pathdir) {
 
     vscode.window.showInputBox({
-        prompt: "name of auth folder",
+        prompt: "Enter name of auth folder",
         placeHolder: "Enter name of create or choose auth folder leave empty for default"
     }).then(function (folderName) {
         if (folderName.length == 0) {
             vscode.window.showInformationMessage("Controllers & models main folder selected.");
-            var new_controller_path = "/application/controllers";
-            var new_model_path = "/application/models";
-        } else {
-            var new_controller_path = "/application/controllers/" + folderName;
-            var new_model_path = "/application/models/" + folderName;
         }
 
         vscode.window.showInputBox({
-            prompt: "name of auth controller/model",
+            prompt: "Enter name of auth controller/model",
             placeHolder: "Enter controller & model name"
         }).then(function (val) {
             if (val.length == 0) {
@@ -48,10 +43,9 @@ module.exports = function (vscode, fs, path, pathdir) {
                             fs.open(modelPath, "w+", function (err, fd) {
                                 if (err) throw err;
                                 fs.writeFileSync(fd, `<?php 
-                
 defined('BASEPATH') OR exit('No direct script access allowed');
-                        
-class ` + capitalize.capitalize(val) + `_model extends CI_Model 
+
+class ${capitalize.capitalize(val)}_model extends CI_Model 
 {
     /*
     You can change users_table to your own.
@@ -103,24 +97,21 @@ class ` + capitalize.capitalize(val) + `_model extends CI_Model
 }
 
 
-/* End of file ` + capitalize.capitalize(val) + `_model.php and path `+new_model_path+`/` + val + `_model.php */
-
-
+/* End of file ${capitalize.capitalize(val)}_model.php and path ${modelPath.replace(pathdir,'')} */
 `);
                                 fs.close(fd);
 
                                 fs.open(controllerPath, "w+", function (err, fd) {
                                     if (err) throw err;
                                     fs.writeFileSync(fd, `<?php 
-        
 defined('BASEPATH') OR exit('No direct script access allowed');
-        
-class ` + capitalize.capitalize(val) + ` extends CI_Controller
+
+class ${capitalize.capitalize(val)} extends CI_Controller
 {
     public function __construct()
     {
         parent::__construct();
-        $this->load->model('` + folderName + `/` + capitalize.capitalize(val) + `_model', '` + val + `');
+        $this->load->model('${folderName}/${capitalize.capitalize(val)}_model', '${val}');
     }
 
     /*
@@ -137,7 +128,7 @@ class ` + capitalize.capitalize(val) + ` extends CI_Controller
                 echo validation_errors();
                 return;
             }
-            if ($query = $this->` + val + `->login($this->input->post('email'))) {
+            if ($query = $this->${val}->login($this->input->post('email'))) {
                 if ($query->num_rows()) {
                     $row = $query->row();
                     if (password_verify($this->input->post('password'), $row->password)) {
@@ -201,7 +192,7 @@ class ` + capitalize.capitalize(val) + ` extends CI_Controller
                 echo validation_errors();
                 return;
             }
-            if ($this->` + val + `->register($this->input->post('email'), password_hash($this->input->post('password'), PASSWORD_BCRYPT))) {
+            if ($this->${val}->register($this->input->post('email'), password_hash($this->input->post('password'), PASSWORD_BCRYPT))) {
 
                 echo 'Account created successful!';
                 return;
@@ -219,9 +210,8 @@ class ` + capitalize.capitalize(val) + ` extends CI_Controller
 }
 
 
-/* End of file ` + capitalize.capitalize(val) + `.php and path `+new_controller_path+`/` + val + `.php */
-        
-                            `);
+/* End of file ${capitalize.capitalize(val)}.php and path ${controllerPath.replace(pathdir,'')}  */
+`);
                                     fs.close(fd);
 
                                     var controllerOpenPath = vscode.Uri.file(controllerPath); //A request file path